line-gateway/requirements.txt
fastapi
uvicorn
line-bot-sdk
httpx
